Fixed image handling with fallback, added checkout flow, converted prices to DecimalField, reorganized media and static folders

+++ flower_shop/flowers/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Flower(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     image = models.ImageField(upload_to="flowers/", null=True, blank=True)
     def __str__(self):
         return self.name
@@ -12,7 +12,7 @@ class Flower(models.Model):
 class Bouquet(models.Model):
     name = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     image = models.ImageField(upload_to="bouquets/", null=True, blank=True)
     flowers = models.ManyToManyField('Flower')
     created_at = models.DateTimeField(auto_now_add=True)
@@ -25,7 +25,7 @@ class Order(models.Model):
     name = models.CharField(max_length=200)  # customer's name
     phone = models.CharField(max_length=15)  # phone number
     address = models.TextField()
-    total_price = models.FloatField()  
+    total_price = models.DecimalField(max_digits=8, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
